create controllers of client and specialist/data fragmentation 1.1
